pyproject.toml
[tool.poetry]
name = "ip-to-ioc"
version = "0.1.0"
description = ""
authors = ["ranguli <hello@joshmurphy.ca>"]

[tool.poetry.dependencies]
python = "^3.6"
OTXv2 = "^1.5.7"
click = "^7.0"
python-decouple = "^3.3"
black = "^19.10b0"
jwt = "^0.6.1"
pytest = "^5.3.5"
sphinx = "^2.4.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
